Block API: try separate controls registration
